refactoring create counter context
