+++ apps/thundermoon_web/lib/thundermoon_web/live/counter_live.ex
@@ -3,7 +3,6 @@ defmodule ThundermoonWeb.CounterLive do
 
   import Canada.Can
 
-  alias Thundermoon.CounterRealm
   alias Thundermoon.Counter
   alias Thundermoon.Repo
   alias Thundermoon.Accounts.User
@@ -14,8 +13,8 @@ defmodule ThundermoonWeb.CounterLive do
   def mount(session, socket) do
     user = Repo.get!(User, session[:current_user_id])
     Endpoint.subscribe("counter")
-    {:ok, _} = CounterRealm.create()
-    digits = CounterRealm.get_digits()
+    {:ok, _} = Counter.create()
+    digits = Counter.get_digits()
     {:ok, assign(socket, current_user: user, digits: digits)}
   end
 
@@ -24,18 +23,18 @@ defmodule ThundermoonWeb.CounterLive do
   end
 
   def handle_event("inc", value, socket) when value in ["1", "10", "100"] do
-    CounterRealm.inc(value)
+    Counter.inc(value)
     {:noreply, socket}
   end
 
   def handle_event("dec", value, socket) when value in ["1", "10", "100"] do
-    CounterRealm.dec(value)
+    Counter.dec(value)
     {:noreply, socket}
   end
 
   def handle_event("reset", _value, socket) do
     if socket.assigns.current_user |> can?(:reset, Counter) do
-      CounterRealm.reset()
+      Counter.reset()
     end
 
     {:noreply, socket}
