+++ apps/thundermoon_web/test/thundermoon_web/live/counter_live_test.exs
@@ -5,7 +5,7 @@ defmodule ThundermoonWeb.CounterLiveTest do
 
   import ThundermoonWeb.AuthSupport
 
-  alias Thundermoon.CounterRealm
+  alias Thundermoon.Counter
 
   def login_as_member(%{conn: conn}) do
     conn = login_as(conn, %{username: "crumb"})
@@ -32,25 +32,25 @@ defmodule ThundermoonWeb.CounterLiveTest do
       assert html_response(conn, 200) =~ "<h1>Counter</h1>"
       {:ok, _view, html} = live(conn)
       assert html =~ "<h1>Counter</h1>"
-      CounterRealm.reset()
+      Counter.reset()
     end
 
     test "inc digit 10", %{conn: conn} do
       @endpoint.subscribe("counter")
       {:ok, view, _html} = live(conn, "/counter")
       render_click(view, :inc, "10")
-      assert_receive(%{event: "update", payload: %{digit_10: 1,}})
-      CounterRealm.reset()
+      assert_receive(%{event: "update", payload: %{digit_10: 1}})
+      Counter.reset()
     end
 
     test "dec digit 10", %{conn: conn} do
       @endpoint.subscribe("counter")
       {:ok, view, _html} = live(conn, "/counter")
       render_click(view, :inc, "10")
-      assert_receive(%{event: "update", payload: %{digit_10: 1,}})
+      assert_receive(%{event: "update", payload: %{digit_10: 1}})
       render_click(view, :dec, "10")
-      assert_receive(%{event: "update", payload: %{digit_10: 0,}})
-      CounterRealm.reset()
+      assert_receive(%{event: "update", payload: %{digit_10: 0}})
+      Counter.reset()
     end
 
     # test "sees users", %{conn: conn} do
